% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_class.R
\name{print.ray_result}
\alias{print.ray_result}
\title{Print a \code{ray_result}}
\usage{
\method{print}{ray_result}(x, ...)
}
\arguments{
\item{x}{A \code{ray_result} object.}

\item{...}{Unused.}
}
\value{
The input object, invisibly.
}
\description{
Print a \code{ray_result}
}
