% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{ray_build_treetools}
\alias{ray_build_treetools}
\title{Build a local TreeTools image}
\usage{
ray_build_treetools(
  context_dir,
  image = ray_default_treetools_image(),
  echo = interactive()
)
}
\arguments{
\item{context_dir}{Docker build context directory.}

\item{image}{Image tag to build.}

\item{echo}{Logical; print command output.}
}
\value{
Invisibly returns Docker build output.
}
\description{
Build a local TreeTools image
}
