% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cog.R
\name{ray_classify_stem}
\alias{ray_classify_stem}
\title{Classify cylinders into trunk vs branch using the highest-tip path}
\usage{
ray_classify_stem(qsm, tree_id_col = "tree_id")
}
\arguments{
\item{qsm}{QSM-like data.frame from ray_treefile_to_qsm()}

\item{tree_id_col}{Name of tree id column}
}
\value{
qsm with added column \code{part}
}
\description{
Classify cylinders into trunk vs branch using the highest-tip path
}
