% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cog.R
\name{ray_compute_cog}
\alias{ray_compute_cog}
\title{Compute centre of gravity for one or more trees}
\usage{
ray_compute_cog(
  qsm,
  density_trunk = 600,
  density_branch = 600,
  branch_factor = 1,
  tree_id_col = "tree_id",
  part_col = "part",
  shift_to_base = TRUE
)
}
\arguments{
\item{qsm}{QSM-like data.frame from ray_treefile_to_qsm()}

\item{density_trunk}{Wood density for trunk cylinders (kg/m^3)}

\item{density_branch}{Wood density for branch cylinders (kg/m^3)}

\item{branch_factor}{Optional multiplier for branch mass}

\item{tree_id_col}{Name of tree id column}

\item{part_col}{Name of trunk/branch classification column}

\item{shift_to_base}{If TRUE, return COG relative to stem base}
}
\value{
data.frame with one row per tree
}
\description{
Compute centre of gravity for one or more trees
}
