% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cog.R
\name{ray_compute_cog_single}
\alias{ray_compute_cog_single}
\title{Compute COG for a single tree}
\usage{
ray_compute_cog_single(qsm_single, ...)
}
\arguments{
\item{qsm_single}{QSM-like data.frame for one tree}

\item{...}{Passed to ray_compute_cog()}
}
\value{
one-row data.frame
}
\description{
Compute COG for a single tree
}
