% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{ray_default_paths}
\alias{ray_default_paths}
\title{Get default project paths}
\usage{
ray_default_paths(project_root = "D:/packageR")
}
\arguments{
\item{project_root}{Root project directory}
}
\value{
Named list of standard directories
}
\description{
Get default project paths
}
