% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{ray_default_raycloud_image}
\alias{ray_default_raycloud_image}
\title{Default RayCloudTools image}
\usage{
ray_default_raycloud_image()
}
\value{
A length-1 character vector with a Docker image tag.
}
\description{
Returns the package default Docker image used for RayCloudTools commands.
}
