% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cog.R
\name{ray_describe_cog}
\alias{ray_describe_cog}
\title{Describe centre-of-gravity output in a compact way}
\usage{
ray_describe_cog(cog_df)
}
\arguments{
\item{cog_df}{Output from ray_compute_cog() or ray_compute_cog_single()}
}
\value{
data.frame with rounded reporting fields
}
\description{
Describe centre-of-gravity output in a compact way
}
