% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{ray_image_has_tools}
\alias{ray_image_has_tools}
\title{Check whether a Docker image exposes required tools}
\usage{
ray_image_has_tools(
  image = ray_default_raycloud_image(),
  tools = c("rayimport", "rayextract", "raysplit", "treesmooth", "treemesh", "treeinfo"),
  echo = FALSE
)
}
\arguments{
\item{image}{Docker image tag.}

\item{tools}{Character vector of command names to check.}

\item{echo}{Logical; show command output while checking.}
}
\value{
A named list of logical values.
}
\description{
Check whether a Docker image exposes required tools
}
