% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ray_plot_cog_compass}
\alias{ray_plot_cog_compass}
\title{Plot COG compass}
\usage{
ray_plot_cog_compass(
  res,
  max_radius = NULL,
  file = NULL,
  width = 800,
  height = 800,
  pad_factor = 1.12,
  label_offset = 0.08
)
}
\arguments{
\item{res}{Output of ray_compute_cog()/ray_compute_cog_single(), or a QSM table}

\item{max_radius}{Optional radius to use for the compass}

\item{file}{Optional output PNG path}

\item{width}{Width in pixels for PNG export}

\item{height}{Height in pixels for PNG export}

\item{pad_factor}{Extra padding around the outer circle}

\item{label_offset}{Relative label offset from the COG point}
}
\value{
Invisibly returns the plotting data.frame
}
\description{
Plot COG compass
}
