% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{ray_pull_image}
\alias{ray_pull_image}
\title{Pull a Docker image}
\usage{
ray_pull_image(image = ray_default_raycloud_image(), echo = interactive())
}
\arguments{
\item{image}{Docker image tag to pull.}

\item{echo}{Logical; print command output.}
}
\value{
Invisibly returns Docker command output.
}
\description{
Pull a Docker image
}
