% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_class.R
\name{ray_result_statuses}
\alias{ray_result_statuses}
\title{Allowed \code{ray_result} statuses}
\usage{
ray_result_statuses()
}
\value{
A character vector of valid status values.
}
\description{
Allowed \code{ray_result} statuses
}
