% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{ray_set_images}
\alias{ray_set_images}
\title{Set package Docker image options}
\usage{
ray_set_images(raycloud_image, treetools_image = NULL)
}
\arguments{
\item{raycloud_image}{Docker image for RayCloudTools commands.}

\item{treetools_image}{Optional Docker image for TreeTools commands.}
}
\value{
Invisibly returns the resolved image list.
}
\description{
Set package Docker image options
}
