% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treefile_to_qsm.R
\name{ray_treefile_to_qsm}
\alias{ray_treefile_to_qsm}
\title{Convert parsed RayExtract treefile/treeinfo data to a QSM-like table}
\usage{
ray_treefile_to_qsm(x)
}
\arguments{
\item{x}{A data.frame returned by ray_read_treefile() or ray_read_treeinfo(),
or a path to a treefile / info file.}
}
\value{
A data.frame with start/end coordinates, geometric segment length,
and direction vectors.
}
\description{
Convert parsed RayExtract treefile/treeinfo data to a QSM-like table
}
