% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_class.R
\name{summary.ray_result}
\alias{summary.ray_result}
\title{Summarize a \code{ray_result}}
\usage{
\method{summary}{ray_result}(object, ...)
}
\arguments{
\item{object}{A \code{ray_result} object.}

\item{...}{Unused.}
}
\value{
A small summary list.
}
\description{
Summarize a \code{ray_result}
}
